==> Rcode/9. JapanSuicide_data_merge.R <==
library(readxl); library(dplyr); library(tidyr)
rm(list=ls())
setwd("C:/Users/masan/Documents/suicide")
load('suicideMen.RDa'); head(data)
suicideMen = data[, c(2, 4, 6:13, ncol(data))]; head(suicideMen)
colnames(suicideMen)[1] = 'pref'; head(suicideMen)
suicideMen = suicideMen %>% pivot_longer(cols = -c('pref', 'year'), 
                                         names_to = "age", values_to = "suicide"); head(suicideMen)
suicideMen = suicideMen %>% mutate(age = ifelse(age == '20歳未満', 20, age), 
                                   age = ifelse(age == '80歳以上', 80, age),
                                   gender = 'male') %>% 
  filter(age != '自殺死亡率' & pref != '不明') %>% mutate(suicide = as.numeric(suicide)); table(suicideMen$age)
unique(suicideMen$pref)
load('suicideWomen.RDa'); head(data)
suicideWomen = data[, c(2, 4, 6:13, ncol(data))]; head(suicideWomen)
colnames(suicideWomen)[1] = 'pref'; head(suicideWomen)
suicideWomen = suicideWomen %>% pivot_longer(cols = -c('pref', 'year'), 
                                             names_to = "age", values_to = "suicide"); head(suicideWomen)
suicideWomen = suicideWomen %>% mutate(age = ifelse(age == '20歳未満', 20, age), 
                                       age = ifelse(age == '80歳以上', 80, age),
                                       gender = 'female') %>% 
  filter(age != '自殺死亡率' & pref != '不明') %>% mutate(suicide = as.numeric(suicide)); table(suicideWomen$age)
aggregate(suicide ~ age, data = suicideMen, mean)
aggregate(suicide ~ age, data = suicideWomen, mean)
aggregate(suicide ~ year, data = suicideMen, mean)
aggregate(suicide ~ year, data = suicideWomen, mean)
subset(suicideWomen, year ==2024 & pref == '全国')
suicide = rbind(suicideMen, suicideWomen); head(suicide); table(suicide$year)
load('pop.RDa'); head(pop)
table(suicide$year); table(pop$year)
pop$age = gsub('0', '0-', pop$age)
pop$age = gsub('1519', '20', pop$age)
table(pop$age); table(suicide$age)
df = merge(pop, suicide, by = c('pref', 'year', 'age', 'gender'))
df$rate = 100000*as.numeric(df$suicide) / df$pop
aggregate(rate ~ age, data = df, mean)
aggregate(rate ~ year, data = df, mean)
aggregate(rate ~ pref, data = df, mean)
aggregate(rate ~ age, data = df[df$gender=='male' & df$pref == '全国',], mean)
aggregate(rate ~ age, data = df[df$gender=='female' & df$pref == '全国',], mean)
load('JapanMW.RDa'); head(data); unique(data$pref); table(data$year)
dataLag1 = data %>% mutate(year = year + 1) %>% rename('lagmw' = 'mw'); head(dataLag1)
dataLag2 = data %>% mutate(year = year + 2) %>% rename('lag2mw' = 'mw'); head(dataLag2)
dataLag = merge(dataLag1, dataLag2, by = c('pref', 'year')); head(dataLag)
data = merge(data, dataLag, by = c('pref', 'year'), all.x = T); tail(data); table(data$year)
table(df$pref); table(data$pref)
df$pref = gsub('県|府', '', df$pref)
df$pref = gsub('東京都', '東京', df$pref)
mw = merge(df, data, by = c('pref', 'year'), all.x = T); table(mw$year); summary(mw)
load('X.RDa'); head(X)
X$pref = gsub('県|府', '', X$pref)
X$pref = gsub('東京都', '東京', X$pref)
unique(X$pref)
data = merge(mw, X, by = c('pref', 'year'), all.x = T); head(subset(data, year == 2024))
data$region = NA
data$region = ifelse(data$pref %in% c("北海道", "青森", "岩手", "宮城", "秋田", "山形", "福島"), "東北", data$region)
data$region = ifelse(data$pref %in% c("埼玉", "千葉", "東京", "神奈川"), "南関東", data$region)
data$region = ifelse(data$pref %in% c("茨城", "栃木", "群馬", "山梨", "長野"), "北関東", data$region)
data$region = ifelse(data$pref %in% c("新潟", "富山", "石川", "福井"), "北陸", data$region)
data$region = ifelse(data$pref %in% c("岐阜", "静岡", "愛知", "三重"), "東海", data$region)
data$region = ifelse(data$pref %in% c("滋賀", "京都", "大阪", "兵庫", "奈良", "和歌山"), "近畿", data$region)
data$region = ifelse(data$pref %in% c("鳥取", "島根", "岡山", "広島", "山口"), "中国", data$region)
data$region = ifelse(data$pref %in% c("徳島", "香川", "愛媛", "高知"), "四国", data$region)
data$region = ifelse(data$pref %in% c("福岡", "佐賀", "長崎", "熊本", "大分", "宮崎", "鹿児島", "沖縄"), "九州", data$region)
table(data$region); subset(data, is.na(region) & year == 2023 & age == '20-29')
summary(data)
load('cancer.RDa'); head(cancer)
cancer$pop = NULL; cancer$fips = NULL
data = merge(data, cancer, all.x = T, by = c('pref', 'year', 'gender', 'age')); head(data)
load('accident.RDa');head(accident)
accident$fips = NULL
data = merge(data, accident, all.x = T, by = c('pref', 'year')); head(data)
save(data, file = 'suicideMW.RDa'); colnames(data); summary(data)
